moved summary list methods from content to utils
